Add files via upload
Wherein I have begun work on the evaluater after realizing that the arrays I was trying to maintain were useless, instead opting for factory methods in the evaluator that will return the arrays when I need them to calculate percentages.  I also attempted to format the strings as they come in through the webReaderFilter to avoid work later but that didn't work.