Add files via upload
Wherein I finally finished dcf after much consternation.  I need to look into and add a math context in such a way that it works properly, or just quit being lazy and do it as I did it and then redo my test results and make it work.  I also refactored dcf so that it's not so much of a mess and added tests for each method and then tests of multiple different values for some of the new tests I added.  I added fairValue and I am in the middle of testing safeValue, which so far looks good.  It looks like data persistence is next.