Add files via upload
Wherein I added a few methods that need to be tested at the bottom of evaluator, including the method that takes stock as a parameter and stores all the relevant data.  I need to continue learning GSON and to delete profile before storing the stock.  That is job 2 after testing the added methods.  I also need to finish the gson video I'm watching.