Add files via upload
Wherein I did a bit of refactoring of the table panels, wherein I added a click event to the popup menu and managed to select and return the values in the selected cell after much consternation.  I also added a descriptions class and added the first description, though I have not yet mapped the description to the column.  Still no data persistence.