Add files via upload
Wherein I wrote the gridbaglayout for multi entry pane, retrieved the information in the form of a hashmap, translated that hashmap to a usable hashmap, and then started trying to deal with zero values.  That is what I am currently doing is trying to deal with zero values.  I also added a hyperlink to the pane.