Add files via upload
Wherein I finished descriptions, tested them, added them to the query event, wrote the resynch event (which works gloriously and was far easier than I thought it would be), removed some fluff code, and celebrated when I realized that really the only thing left to do is add Json.  I also added the update method.  Implementation was different than I thought it would be.  Instead of checking the date I just query the API each time the program is opened and update those that are set to auto update.